Fix rbenv role path issues
The rbenv role modifies the user's path here (by default) by inserting an additional GEM_PATH in the user's environment but it's not actually the one we use. In most cases it gets ignored anyway, but it can cause gems to not be found in some cases when processes fork.